+++ doc/ChangeLog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # master
 
+* [#99](https://github.com/BlueBrain/Brayns/pull/99):
+  Add livre as a new rendering engine
 * [#95](https://github.com/BlueBrain/Brayns/pull/95):
   Upgraded to OSPRay 1.1.2
 * [#87](https://github.com/BlueBrain/Brayns/pull/87):